HF_Space_gemma2_2b_texttosql/requirements.txt
gguf>=0.10.0
streamlit
llama-cpp-python
